feat: Upgrade Spring Boot v4.0 and Spring Modulith v2.0

+++ spring-boot-starter-event-system/src/test/java/fr/fezlight/eventsystem/EventServiceIT.java
@@ -49,10 +49,10 @@ public class EventServiceIT {
 
         eventService.reprocessAllFailedMessage();
 
-        int countError = amqpAdmin.getQueueInfo("events.error").getMessageCount();
-        assertThat(countError).isEqualTo(0);
-        Integer count = amqpAdmin.getQueueInfo("events.testevents.worker").getMessageCount();
-        assertThat(count).isEqualTo(1);
+        Long countError = amqpAdmin.getQueueInfo("events.error").getMessageCount();
+        assertThat(countError).isEqualTo(0L);
+        Long count = amqpAdmin.getQueueInfo("events.testevents.worker").getMessageCount();
+        assertThat(count).isEqualTo(1L);
     }
 
     public record TestEventService(String eventName) implements Event {
